chore: move ExampleSuite logic to codegen core so it can be used for Scala as well
